% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSubcellulaRvis.R
\name{compartmentData}
\alias{compartmentData}
\title{Calculate enrichment of subcellular compartment}
\usage{
compartmentData(genes, bkgdSize = 20367, trafficking = F, subAnnots = F)
}
\arguments{
\item{genes}{input genes (HGNC symbol) as character vector}

\item{bkgdSize}{size of background proteome. Default = 20367, HeLa proteome}

\item{trafficking}{Logical; Calculate enrichment within endolysosome system?}
}
\value{

}
\description{
Calculate enrichment of subcellular compartment
}
\examples{

}
