% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mouse_exampleGenes}
\alias{Mouse_exampleGenes}
\title{Mouse gene list to be used as example input}
\format{
A data frame with 345 rows and 1 variables:
\describe{
\item{v1}{Mouse genes}
...
}
}
\usage{
Mouse_exampleGenes
}
\description{
Mouse gene list to be used as example input
}
\keyword{datasets}
