% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSubcellulaRvis.R
\name{plotOverlap}
\alias{plotOverlap}
\title{Plot Upset plot of compartment annotations overlap of genes}
\usage{
plotOverlap(compsDat)
}
\arguments{
\item{compsDat}{Dataframe output from compartmentData}
}
\value{
list containing upset plot and associated data
}
\description{
Plot Upset plot of compartment annotations overlap of genes
}
