% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSubcellulaRvis.R
\name{runSubcellulaRvis}
\alias{runSubcellulaRvis}
\title{Visualise subcellular enrichment}
\usage{
runSubcellulaRvis(
  compsDat,
  colScheme_low,
  colScheme_high,
  trafficking = F,
  text_size = 2,
  legend = T,
  legend.pos = "right",
  labels = T
)
}
\arguments{
\item{compsDat}{Output of compartmentData()}

\item{colScheme_low}{Low value (i.e. most statistically ignificant) of colour scheme}

\item{colScheme_high}{High value (i.e. most statistically significant) of colour scheme}

\item{trafficking}{Logical; Visualise endolysosome system?}

\item{text_size}{Size of text if using}

\item{legend}{Logical; include legend?}

\item{legend.pos}{passes to ggplot2::theme(legend.postion = legend.pos). One of "right", "left", "bottom", "top"}
}
\value{

}
\description{
Visualise subcellular enrichment
}
