% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellularapp.R
\name{subcellularapp}
\alias{subcellularapp}
\title{Run Shiny App for SubcellularVis}
\usage{
subcellularapp(...)
}
\arguments{
\item{...}{empty}
}
\value{
Shiny App
}
\description{
Run Shiny App for SubcellularVis
}
