% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSubcellulaRvis.R
\name{compartmentData}
\alias{compartmentData}
\title{Calculate enrichment of subcellular compartment}
\usage{
compartmentData(
  genes,
  bkgd = NULL,
  id_type = c("SYMBOL", "UNIPROT"),
  aspect = c("Whole cell", "Endosomal systems"),
  subAnnots = F,
  organism = c("Human", "Mouse", "Drosophila", "Yeast", "Rat", "Xenopus"),
  annotationSource = c("Gene Ontology", "Human Protein Atlas"),
  significanceThresh = 0.05
)
}
\arguments{
\item{genes}{input genes (HGNC symbol) as character vector}

\item{bkgd}{background population. Defaults to size of species library}

\item{id_type}{Character; either "SYMBOL" or "UNIPROT", relating to identifier type of gene list.}

\item{aspect}{Character; either "Whole cell" or "Endosomal system.
Latter will calculate enrichment within endosomal system. system?}

\item{subAnnots}{Logical; Calculate enrichment on all terms, not summarised}

\item{organism}{One of "Human", "Mouse",  "Drosophila", "Yeast", "Rat", "Xenopus"}

\item{annotationSource}{Annotation source, Gene Ontology or (if organism = "Human") the Human Protein Atlas}

\item{significanceThresh}{Threshold for FDR statistical significance. Default is 0.05}
}
\value{
data frame
}
\description{
Calculate enrichment of subcellular compartment
}
