% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSubcellulaRvis.R
\name{compartmentData}
\alias{compartmentData}
\title{Calculate enrichment of subcellular compartment}
\usage{
compartmentData(
  genes,
  bkgd = NULL,
  trafficking = F,
  subAnnots = F,
  organism = c("Human", "Mouse", "Drosophila", "Yeast", "Rat", "Xenopus")
)
}
\arguments{
\item{genes}{input genes (HGNC symbol) as character vector}

\item{bkgd}{background population. Defaults to size of species library}

\item{trafficking}{Logical; Calculate enrichment within endolysosome system?}

\item{subAnnots}{Logical; Calculate enrichment on all terms, not summarised}

\item{organism}{One of "Human", "Mouse",  "Drosophila", "Yeast", "Rat", "Xenopus"}
}
\value{

}
\description{
Calculate enrichment of subcellular compartment
}
