% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Human_exampleGenes}
\alias{Human_exampleGenes}
\title{Human gene list to be used as example input}
\format{
A data frame with 346 rows and 1 variables:
\describe{
\item{v1}{Human genes}
...
}
}
\usage{
Human_exampleGenes
}
\description{
Human gene list to be used as example input
}
\keyword{datasets}
